+++ test/foundry/helpers/Skipper.sol
@@ -4,7 +4,7 @@ pragma solidity ^0.8.15;
 import "forge-std/Test.sol";
 
 contract Skipper is Test {
-    function doSkip(uint time) external {
+    function doSkip(uint256 time) external {
         time = bound(time, 1, 3650 days);
         skip(time);
     }
